requirements.txt
python-dotenv
pandas
pydantic
streamlit
requests
rdflib
owlready2
networkx
matplotlib
plotly
openpyxl
google-genai
